[markdown] updates less often, cursor moves
I was getting mad slowdown when I put some tables into my document, so I changed it to no longer update with each modification to the main document. Only on initially opening the dock widget, and obviously on printing.

The cursor updating is a bit wonky, but I like it better than *not* having it, so in it goes for now.